update education level

+++ backend/src/main/java/com/logumrh/repository/common/EducationLevelRepository.java
@@ -3,5 +3,5 @@ package com.logumrh.repository.common;
 import com.logumrh.model.common.EducationLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EducationLevelRepository extends JpaRepository<EducationLevel, Integer> {
+public interface EducationLevelRepository extends JpaRepository<EducationLevel, Long> {
 }
